Update dlipower.py
If the switch does not let you have access to all ports and the port you do have access to has an index larger than the number of ports the switch let's you see, it will throw an exception.

Example: 
I had a switch that only gave me port 5.  Even when specifying the port directly, the script would bomb out because it believed i was asking for a port out of range. Doing the explicit compare on port number avoids this issue.